Modify resource control arbitrator
Sort of tested version of handling resource control using a service/plugin and
a corresponding arbitrator. Compiles, connects and seems to work.

Should be tested further and later modified to allow for more flexible resource
allocation. The latter may require some interface modifications.
Doxygen docs should be expanded (not enough documentation right now).

gait_walking is a mockup of a walking controller-component. It is here only as
an example of using the resource control service (and to test it).